userId changed from Long to id

+++ Application/machine-rental/src/main/java/rent/tycoon/persistance/config/security/token/impl/AccessTokenImpl.java
@@ -12,10 +12,10 @@ import java.util.Set;
 @Getter
 public class AccessTokenImpl implements AccessToken {
     private final String subject;
-    private final Long userId;
+    private final int userId;
     private final Set<String> roles;
 
-    public AccessTokenImpl(String subject, Long userId, Collection<String> roles) {
+    public AccessTokenImpl(String subject, int userId, Collection<String> roles) {
         this.subject = subject;
         this.userId = userId;
         this.roles = roles != null ? Set.copyOf(roles) : Collections.emptySet();
